Try and use webuntis teachers' from teachers.json
We don't have acces to get teachers from within the API, but we do in
webuntis. If you (manually) save the json response from webuntis and
put it in teachers.json; this file will be used to resolve teacher id's.
Better would be to have a script download this. Even better would be to
integrate it in the API.